move more things under v1 namespace
